requirements.txt
argparse
os
tarfile
collections
gzip
multiprocessing
pandas
biopython